get current keybox from phone and show its info

+++ scripts/install_android.sh
@@ -25,7 +25,7 @@ get_keybox_path() {
 if [ -e "$KEY_FILE" ]; then
   get_keybox_path
 
-  if [ -n "$KEY_BOX" ]; then
+  if [ -n "$KEY_BOX" ] && [ -e "$KEY_BOX" ]; then
     cp "$KEY_FILE" "$KEY_BOX"
     chmod 644 "$KEY_BOX"
     chown root:root "$KEY_BOX"
